+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 0.1.2
+
+- Fix bug where Loggly instance can only be created once
+- Add 'module' name property to logged object
+
 ### 0.1.1
 
 - Don't override errorId if it's passed in as part of the log object.